+++ src/serverobject.h
@@ -188,7 +188,7 @@ public:
 	{}
 	virtual std::string getWieldList() const
 	{ return ""; }
-	virtual int getWieldIndex() const
+	virtual u16 getWieldIndex() const
 	{ return 0; }
 	virtual ItemStack getWieldedItem() const;
 	virtual bool setWieldedItem(const ItemStack &item);
